+++ tests/src/Opulence/Orm/Ids/Accessors/Mocks/Foo.php
@@ -1,0 +1,34 @@
+<?php
+/**
+ * Opulence
+ *
+ * @link      https://www.opulencephp.com
+ * @copyright Copyright (C) 2015 David Young
+ * @license   https://github.com/opulencephp/Opulence/blob/master/LICENSE.md
+ */
+namespace Opulence\Tests\Orm\Ids\Accessors\Mocks;
+
+/**
+ * Mocks a class with an Id property
+ */
+class Foo
+{
+    /** @var int The Id */
+    private $id = -1;
+
+    /**
+     * @return int
+     */
+    public function getId()
+    {
+        return $this->id;
+    }
+
+    /**
+     * @param int $id
+     */
+    public function setId($id)
+    {
+        $this->id = $id;
+    }
+}